requirements.txt
# YOLOv8 화재 및 연기 감지 프로젝트 필수 패키지
# Python 3.8+ 권장

# 핵심 딥러닝 및 YOLO
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.20.0

# Roboflow 데이터셋 다운로드
roboflow>=1.1.0

# 이미지 처리
opencv-python>=4.5.0
Pillow>=8.0.0
imageio>=2.9.0

# 데이터 처리 및 분석
pandas>=1.3.0
scipy>=1.7.0

# 시각화
matplotlib>=3.3.0
seaborn>=0.11.0
plotly>=5.0.0

# 설정 및 유틸리티
pyyaml>=5.4.0
tqdm>=4.60.0
requests>=2.25.0

# Google Colab 환경용
google-colab  # Colab에서만 자동 설치됨

# 선택적 의존성 (성능 향상용)
# GPU 가속
# tensorrt>=8.0.0  # NVIDIA GPU 환경에서만
# onnx>=1.12.0     # ONNX 모델 내보내기용
# openvino>=2022.1 # Intel OpenVINO용

# 개발 및 테스트용 (선택사항)
# jupyter>=1.0.0
# notebook>=6.4.0
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0 Google Colab 환경용
google-colab  # Colab에서만 자동 설치됨

# 선택적 의존성 (성능 향상용)
# GPU 가속
# tensorrt>=8.0.0  # NVIDIA GPU 환경에서만
# onnx>=1.12.0     # ONNX 모델 내보내기용
# openvino>=2022.1 # Intel OpenVINO용

# 개발 및 테스트용 (선택사항)
# jupyter>=1.0.0
# notebook>=6.4.0
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0